Add PHPUnit tests
